Changed the way configuration works.
The new "ConfigurationSymbolProvider" is now a central possibility to
activate file-delivered configuration for all projects using the
semedico core. This symbol provider looks up a classpath resource by
either the system property semedico.configuration or a configuration
file named configuration.properties.<username>. This is actually the way
it also worked before but encapsuled in a single class which may
contributed to the Tapestry symbol provider.