Further refined configuration file handling.
Configuration files are first looked up on the classpath as they were in
the past. If this is unsuccessful, the file is looked up as an actual
file. This means that you now can reference any configuration file when
starting a semedico application, it doesn't have to be on the classpath.