Reinstated index page layout before bootstrap adaptations.
At the current state, the bootstrap version doesn't look right but we
need some working semedico version with REST api right now. Thus, this
branch is only here to correct the design. It should not be merged back
because the actual goal is to get rid of the self-created CSS as far as
possible and rather use bootstrap.