Fixed suggestions mapping to have ES 5.4 accept it.
With ES 5, the payload feature of suggestions has been dropped in favor
of returning the _source field for suggestions. The semedico code will
have to be adapted for this, but this mapping is accepted by ES (and the
facets field type had to be set to keyword since string is deprecated).