Working on and realizing multi document type, multi document part search.
Added "Broadcasting". The idea is that a single user query is
broadcast onto a selection of document types and parts. The challenge
here is to not only send the query but also collect facets and
the document results from the different document part indexes.
The current idea is to send broadcast objects (found in the
search.broadcasting package) into a chain-of-command where
the "DocModQueryService", which is yet to be created, is made of a
sequence of concrete document part DocModQueryServices that will return
the expected AggregationRequest or SearchResultCollector for the
adequate broadcasts.