Intermediate commit at a blocking issue.
The issue being that the document modules need to reference
the IQuerySelector interface. This interface has depenencies to
other classes that are part of core semedico or belong to the
general search infrastructure of semedico. It will probably be
necessary to create a maven module for the semedico search
capabilities.