Frontend: Cropped everything that caused a compile error.
A lot has changed in the core which the frontend was not adapted to yet.
The most notable change is the removal of the StatefulSearchService.
The core SearchService does not require state any more because its
main principle has been minimalized.
Import not note is that this version of the frontend is far from
functional. Especially all places where a search was issued have
just been commented out or removed alltogether.