Finished first version of broadcast service. Removed SearchScopes.
The QueryBroadcastingService now has a first test that works. It is
a rather simple tests with two document parts queried.
The search scopes have partly been replaced by the document parts which are
a much more concrete way to give the instruction about what to
search and do. What document parts lack in expression should be
complemented by the search options when issuing an actual search.