Created the "SemedicoIndexField", "DefaultQueryTranslator" and "QueryTranslation" classes.
The idea is that most fields in Semedico just have concept IDs in
them or not and are standard text fields aside of that. So a lot of
fields would nicely served by a default query translator that
just knows if there can be concept IDs in the searched field or not.
Perhaps other configuration options in the future. The configuration
options are properties of the new "SemedicoIndexField" field.
Actually, options like "min_must_match" and "wildcard allowed" could
- and should? - natural additions to the properties of the new field
class. Thus, a field just determines how it is searched. And the
respective DocumentPart defines the correct field. Everyone's happy.
Why didn't I come up with this earlier?

Also, removed deprecated classes, renamed method names (term->concept)
and did quite some cleaning.