Added highlighting.
ElasticQuery (sub)classes now directly contain a HighlightCommand.
This is just passed to the request creation component.
Also, all ISemedicoQuery classes can now reveal the type of result
they query, articles or SERP items or unspecified. This
information is used in the QueryBroadcastService which passes it
to the DocModQueryServices. Those can then create the adequate
highlighting.
Search with document modules is still untested, however, only
the core SearchService has been tested so far.