Organized all tests in TestNG suites. All tests running.
There are tests that require Neo4j or ElasticSearch which are
provided through testcontainers. Those need a while to start up
and also the the ES container does not seem to propertly start up
when other tests (the neo4j tests?) are run.
There are now 3 suites for ES, Neo4j and the rest ("general").
The suite XML files are found in the test resources `suites/`
directory. A quick way to run exactly those suites is to
execute `mvn test`, the Surefire plugin has been configured
for the suites. Unfortunately, just running "all tests" of `semedico-core`
in an IDE does not work for me, the suites must be run individually.