Updated the retrieval of all index terms for suggestions.
For this, I created the new ElasticSearchQuery so one can easily
just query all documents. However, this query will most probably
not be accepted in downstream components at the moment, which all
expect parse trees. As a next thing, a test for getting all the
suggestion terms should be added.