Completed an implementation of `LexerService`, untested.
This relates to the scicopia LexerService which should replace
the original Semedico LexerService which was based on JFLex.
The huge advantage of scicopia is the use of ANTLR which allows
for smaller grammars and, very importantly, comes with its own
parser formalism which should replace the old ParsingService and,
thus, the old ParseTree.