Forged the grammar with nonterminals for boolean operations.
The idea was to distinguish the operations very clearly for a
simple as possible traversion through the resulting parse tree.
I.e. if all the contexts are just "QueryContext" one has to apply
quite some logic to find out whether there was an "AND" in the query
or a "OR". However, I just found out about labels. Those allow exactly
to specify which rule should get which enter/exit method in the
listener. Thus, I will try to go back to the more simple grammar
I had before and just give labels.