Fixed the preanalyzed plugin and the test document. Indexing is working.
The plugin needed to have a slightly different internal structure,
see commit messages for the preanalyzed plugin. The preanalyzed document
needed to be restructures in a way that the preanalyzed field value
is actually one single value - a string encoding the preanalyzed json
format - instead of being a complete json sub-structure.