Repairing Semedico-Frontend with Bootstrap 4.
To use Bootstrap 4, Tapestry was updated to `v5.5.0-beta-3` which
supports Bootstrap 4. Everything was already configured to updates
the exact Bootstrap version to not be dependent on the Tapestry
delivered defaults.

The Index.tml page was simplified regarding the `col` class assignments
(no more col-md-12 where not really needed).

Currently, the web resources are restructured to go into the
`META-INF/assets` folder. Deprecated assets will be removed.